Added db_helper_test file, readded DIRPATH and ENGINEPATH
